Move driver script to its own file
This makes it easier to run in IDEs
